customer,product,basket,order ve payment servis yazıldı. Müşteriler kayıt olup sipariş verip ödeme yapabiliyorlar. Iyzico entegresi impl edildi.

+++ product-service/src/main/java/com/os/product_service/controller/CustomerController.java
@@ -23,7 +23,7 @@ public class CustomerController {
 
     // ID'ye göre müşteri getirme
     @GetMapping("/{id}")
-    public ResponseEntity<Customer> getCustomerById(@PathVariable int id) {
+    public ResponseEntity<Customer> getCustomerById(@PathVariable Long id) {
         Customer customer = customerService.getCustomerById(id);
         if (customer != null) {
             return ResponseEntity.ok(customer);
