+++ product-service/src/main/java/com/os/product_service/service/CustomerService.java
@@ -4,5 +4,5 @@ import com.os.product_service.model.Customer;
 
 public interface CustomerService {
     Customer addCustomer(Customer customer);
-    Customer getCustomerById(int id);
+    Customer getCustomerById(Long id);
 }
