+++ product-service/src/main/java/com/os/product_service/service/PaymentService.java
@@ -4,5 +4,5 @@ import com.os.product_service.model.PaymentRequest;
 
 public interface PaymentService {
 
-    String makePayment(int productId, int customerId, PaymentRequest paymentRequest);
+    String makePayment(Long productId, Long customerId, PaymentRequest paymentRequest);
 }
