+++ product-service/src/main/java/com/os/product_service/service/impl/CustomerServiceImpl.java
@@ -21,7 +21,7 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public Customer getCustomerById(int id) {
+    public Customer getCustomerById(Long id) {
         return customerRepository.findById(id).orElse(null);
     }
 }
